Location Additions & Adjustments
Prison and GolfCourse now extend Location as default Locations.

When a location is created, it generates its inventory using an array of numbers, which are passed to a RNG to give the Location a starting inventory.

Structures are created in the Location constructor by passing the keys of the structures that should be created.

Helper.js contains helper methods such as the RNG and temporarily contains the object templates for stats, structures, and inventory. This is to keep these items in one place and only change one file when adding or removing attributes.